+++ packages/google_maps_flutter/google_maps_flutter/example/test_driver/google_map_inspector.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:typed_data';
 import 'package:flutter/services.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
@@ -62,5 +63,9 @@ class GoogleMapInspector {
 
   Future<bool> isBuildingsEnabled() async {
     return await _channel.invokeMethod<bool>('map#isBuildingsEnabled');
+  }
+
+  Future<Uint8List> takeSnapshot() async {
+    return await _channel.invokeMethod<Uint8List>('map#takeSnapshot');
   }
 }