+++ packages/android_intent/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.7
+
+* Add a `Future<bool> canResolveActivity` method to the AndroidIntent class. It 
+  can be used to determine whether a device supports a particular intent or has 
+  an app installed that can resolve it. It is based on PackageManager
+  [resolveActivity](https://developer.android.com/reference/android/content/pm/PackageManager#resolveActivity(android.content.Intent,%20int)).
+
 ## 0.3.6+1
 
 * Bump the minimum Flutter version to 1.12.13+hotfix.5.